Refactor code to fix phpstan errors

+++ src/Controller/FavoriteController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Favorite;
 use App\Entity\Product;
+use App\Entity\User;
 use App\Repository\FavoriteRepository;
 use DateTime;
 use Doctrine\ORM\EntityManagerInterface;
@@ -14,14 +15,29 @@ use Symfony\Component\Routing\Attribute\Route;
 final class FavoriteController extends AbstractController
 {
     #[Route('/favoris', name: 'app_favorite')]
+
     public function index(FavoriteRepository $favoriteRepository): Response
     {
-        $favorites = $favoriteRepository->findFavoritesByUser($this->getUser());
+        /** @var User|null $user */
+        $user = $this->getUser();
+
+        if ($user === null) {
+            return $this->redirectToRoute('app_login');
+        }
+
+        $favorites = $favoriteRepository->findFavoritesByUser($user);
+
         return $this->render('pages/favorite/index.html.twig', [
             'favorites' => $favorites,
         ]);
     }
 
+    /**
+     * @param Product $product
+     * @param EntityManagerInterface $entityManager
+     * @param FavoriteRepository $favoriteRepository
+     * @return Response
+     */
     #[Route('/favoris/ajout/{id}', name: 'app_favorite_add')]
     public function add(
         Product $product,
@@ -29,7 +45,13 @@ final class FavoriteController extends AbstractController
         FavoriteRepository $favoriteRepository,
     ): Response
     {
+        /** @var User|null $user */
         $user = $this->getUser();
+
+        if ($user === null) {
+            return $this->redirectToRoute('app_login');
+        }
+
         $existingFavorite = $favoriteRepository->findOneBy([
             'userAccount' => $user,
             'product' => $product,
@@ -41,7 +63,7 @@ final class FavoriteController extends AbstractController
         }
 
         $favorite = new Favorite();
-        $favorite->setUserAccount($this->getUser());
+        $favorite->setUserAccount($user);
         $favorite->setProduct($product);
         $favorite->setCreatedAt(new DateTime());
 
@@ -71,17 +93,22 @@ final class FavoriteController extends AbstractController
         FavoriteRepository $favoriteRepository
     ): Response
     {
+        /** @var User|null $user */
         $user = $this->getUser();
+
 
         $existingFavorite = $favoriteRepository->findOneBy([
             'userAccount' => $user,
             'product' => $product,
         ]);
 
-
-        $entityManager->remove($existingFavorite);
-        $entityManager->flush();
-        $this->addFlash('success', 'Le produit a bien été retiré de vos favoris');
+        if ($existingFavorite !== null) {
+            $entityManager->remove($existingFavorite);
+            $entityManager->flush();
+            $this->addFlash('success', 'Le produit a bien été retiré de vos favoris');
+        } else {
+            $this->addFlash('warning', 'Ce produit n\'est pas dans vos favoris');
+        }
 
         return $this->redirectToRoute('app_product_show', ['slug' => $product->getSlug()]);
 
@@ -99,8 +126,6 @@ final class FavoriteController extends AbstractController
         EntityManagerInterface $entityManager,
     ): Response
     {
-        $user = $this->getUser();
-
         $entityManager->remove($favorite);
         $entityManager->flush();
         $this->addFlash('success', 'Le produit a bien été retiré de vos favoris');
